+++ src/sentry/seer/explorer/client_utils.py
@@ -208,6 +208,15 @@ def collect_user_org_context(
     try:
         member = OrganizationMember.objects.get(organization=organization, user_id=user.id)
     except OrganizationMember.DoesNotExist:
+        # User is not a member of this organization (e.g., superuser accessing foreign org)
+        logger.warning(
+            "User attempted to access Seer Explorer for organization they are not a member of",
+            extra={
+                "user_id": user.id,
+                "organization_id": organization.id,
+                "organization_slug": organization.slug,
+            },
+        )
         return {
             "org_slug": organization.slug,
             "all_org_projects": all_org_projects,
